Fix PLL config and speed up to 48MHz.
PLL config was all wrong, sending 8MHz to the PLL instead of 4MHz, which ran the PLL at 160MHz which is above the maximum of 120MHz. Also, the chip can run at up to 50MHz when below 85°C. The limit of 40MHz only applies if the chip is hotter, up to 105°C. A little bit more speed here makes the scope display go from 22ms to 18ms and the lines go from 145ms to 121ms.